"the best solution to a problem is usually the easiest one" - GladOS
The current solution is really simple:
First we organize the number in base by their length, then, foreach created group, we create sequences that contains numbers that follow the 1 digit diference, however, we do that foreach and every number from the group.

After that, we just return the sequence with the most numbers.